Add addons to the base Matlab installation
The current addons are:

* NeurodataWithoutBorders v2.6.0.0
* Brain-Observatory-Toolbox v0.9.2

They are downloaded from the github release page of each projects.

Currenlty, the addition of the addons to Matlab is done through a patch
of the "startup.m" script.
This script is automatically loaded by matlab-proxy when an instance
of Matlab is launched.
The script is located here:
"/opt/conda/lib/python3.10/site-packages/matlab_proxy/matlab/startup.m"