Build and (optionally, only in main) push docker images in a single workflow
Duplication is evil!  So, this is what claude code came up with and I just
added also reacting to changes to the actual workflow file as well.
I had to do that in https://github.com/dandi/dandi-hub/pull/254 which ATM
tests building against older ubuntu 22.04 which seems to not run out of the space